pyproject.toml
[tool.poetry]
name = "langraph-agentic-systems"
version = "0.1.0"
description = "Building Agentic systems using Langraph"
authors = ["Pragnesh Anekal <72695068+Pragnesh260@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
docling = "^2.4.2"
langchain-text-splitters = "^0.3.2"
langchain-huggingface = "^0.1.2"
langchain-pinecone = "^0.2.0"
langchain-openai = "^0.2.6"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
langgraph = "^0.2.45"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
